[pd_controller] Rename files/dir according to the naming convention
